requirements.txt
google-cloud-ndb
flask~=1.1.2